move SQL Alchemy ORM models to the importer module
