+++ soweego/importer/models/base_link_entity.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-"""TODO module docstring"""
+"""Base SQL Alchemy ORM link entity"""
 
 __author__ = 'Marco Fossati'
 __email__ = 'fossati@spaziodati.eu'
@@ -10,14 +10,12 @@ __license__ = 'GPL-3.0'
 __copyright__ = 'Copyleft 2018, Hjfocs'
 
 
-from sqlalchemy import Column, ForeignKey, Index, Integer, String
-from sqlalchemy.engine import Engine
+from sqlalchemy import Column, Index, Integer, String
+
 
 class BaseLinkEntity(object):
-
     internal_id = Column(Integer, unique=True,
                          primary_key=True, autoincrement=True)
-
     # Full URL
     url = Column(String(255))
     # Tokenized URL
