+++ soweego/importer/models/discogs_musician_entity.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-"""Discogs ORM model for musicians"""
+"""Discogs SQL Alchemy ORM model for musicians"""
 
 __author__ = 'Marco Fossati'
 __email__ = 'fossati@spaziodati.eu'
@@ -9,25 +9,22 @@ __version__ = '1.0'
 __license__ = 'GPL-3.0'
 __copyright__ = 'Copyleft 2018, Hjfocs'
 
-from sqlalchemy import Column, ForeignKey, Integer, String, Text
+from sqlalchemy import Column, ForeignKey, String, Text
 from sqlalchemy.ext.declarative import declarative_base
 
-from soweego.commons.models.base_entity import BaseEntity
-from soweego.commons.models.base_link_entity import BaseLinkEntity
+from soweego.importer.models.base_entity import BaseEntity
+from soweego.importer.models.base_link_entity import BaseLinkEntity
 
-Base = declarative_base()
-
+BASE = declarative_base()
 DISCOGS_MUSICIAN_TABLE_NAME = 'discogs_musician'
 DISCOGS_MUSICIAN_LINK_TABLE_NAME = 'discogs_musician_link'
 
 
-class DiscogsMusicianEntity(BaseEntity, Base):
+class DiscogsMusicianEntity(BaseEntity, BASE):
     __tablename__ = DISCOGS_MUSICIAN_TABLE_NAME
-    internal_id = Column(Integer, ForeignKey(BaseEntity.internal_id),
-                         primary_key=True, autoincrement=True)
     # Discogs identifier of a group the musician belongs to
     group_id = Column(String, ForeignKey(
-        '%s.catalog_id' % DISCOGS_MUSICIAN_TABLE_NAME), index=True) #TODO find a better way 
+        '%s.catalog_id' % DISCOGS_MUSICIAN_TABLE_NAME), index=True)  # TODO find a better way
     # Name in real life
     real_name = Column(String)
     # Other art names
@@ -40,7 +37,7 @@ class DiscogsMusicianEntity(BaseEntity, Base):
     data_quality = Column(String)
 
 
-class DiscogsMusicianLinkEntity(BaseLinkEntity, Base):
+class DiscogsMusicianLinkEntity(BaseLinkEntity, BASE):
     __tablename__ = DISCOGS_MUSICIAN_LINK_TABLE_NAME
     catalog_id = Column(String(32), ForeignKey(DiscogsMusicianEntity.catalog_id),
                         index=True)
